fix: satisfy max-lines ratchet
