+++ test/widget/alera_shell_page_test.dart
@@ -38,6 +38,7 @@ import 'package:flutter_test/flutter_test.dart';
 part 'alera_shell_page_test_harness.dart';
 part 'alera_shell_page_runtime_test_harness.dart';
 part 'alera_shell_page_workbench_test_cases.dart';
+part 'alera_shell_page_codex_sidebar_test_cases.dart';
 part 'alera_shell_page_sidebar_actions_test_cases.dart';
 part 'alera_shell_page_sidebar_states_test_cases.dart';
 part 'alera_shell_page_pinning_test_cases.dart';
@@ -111,6 +112,7 @@ class _ShellAgentProfiles extends AgentProfiles {
 
 void main() {
   _registerAleraShellWorkbenchTests();
+  _registerAleraShellCodexSidebarTests();
   _registerAleraShellSidebarActionTests();
   _registerAleraShellSidebarStateTests();
   _registerAleraShellPinningTests();
